+++ src/main/java/com/gmeister/temp/pkcmmsrando/Notes.java
@@ -26,7 +26,6 @@ import com.gmeister.temp.pkcmmsrando.map.data.Player;
 import com.gmeister.temp.pkcmmsrando.map.data.TileSet;
 import com.gmeister.temp.pkcmmsrando.map.data.Warp;
 import com.gmeister.temp.pkcmmsrando.rando.Randomiser;
-import com.gmeister.temp.pkcmmsrando.rando.WarpRandomiser;
 
 public class Notes
 {
@@ -116,7 +115,7 @@ public class Notes
 		}
 	}
 	
-	public static void randomiseWarpAreas(List<Map> maps, EmpiricalDataReader empReader, WarpRandomiser rando) throws IOException, URISyntaxException
+	public static void randomiseWarpAreas(List<Map> maps, EmpiricalDataReader empReader, Randomiser rando) throws IOException, URISyntaxException
 	{
 		List<List<Warp>> warpGroups = new ArrayList<>();
 		List<List<Map>> mapGroups = new ArrayList<>();
@@ -150,7 +149,7 @@ public class Notes
 		rando.shuffleWarpGroups(warpGroups, false, true);
 	}
 	
-	public static void buildWarpAreas(List<Map> maps, List<Flag> flags, EmpiricalDataReader empReader, WarpRandomiser rando) throws FileNotFoundException, IOException, URISyntaxException
+	public static void buildWarpAreas(List<Map> maps, List<Flag> flags, EmpiricalDataReader empReader, Randomiser rando) throws FileNotFoundException, IOException, URISyntaxException
 	{
 		List<List<Warp>> warpGroups = new ArrayList<>();
 		List<List<Map>> mapGroups = new ArrayList<>();
@@ -283,7 +282,7 @@ public class Notes
 		return selectedMaps;
 	}
 	
-	public static void randomiseWarps(ArrayList<Map> maps, WarpRandomiser rando) throws IOException
+	public static void randomiseWarps(ArrayList<Map> maps, Randomiser rando) throws IOException
 	{
 		//Collect a bunch of maps to manually edit warps
 		Map victoryRoadGate = maps.stream().filter(m -> m.getConstName().equals("VICTORY_ROAD_GATE")).findFirst().orElseThrow();
@@ -517,7 +516,6 @@ public class Notes
 		
 		EmpiricalDataReader empReader = new EmpiricalDataReader(null);
 		Randomiser rando = new Randomiser();
-		WarpRandomiser warpRando = new WarpRandomiser();
 		Disassembly disassembly = new Disassembly();
 		ArrayList<Flag> allFlags = new ArrayList<>();
 		
@@ -539,7 +537,7 @@ public class Notes
 		else if (warps)
 		{
 			if (disReader == null) System.out.println("Error: Randomisers require -d");
-			Notes.randomiseWarps(disassembly.getMaps(), warpRando);
+			Notes.randomiseWarps(disassembly.getMaps(), rando);
 			
 			if (disWriter != null) for (Map map : disassembly.getMaps())
 			{
@@ -550,7 +548,7 @@ public class Notes
 		else if (warpAreas)
 		{
 			if (disReader == null) System.out.println("Error: Randomisers require -d");
-			Notes.buildWarpAreas(disassembly.getMaps(), allFlags, empReader, warpRando);
+			Notes.buildWarpAreas(disassembly.getMaps(), allFlags, empReader, rando);
 			
 			if (disWriter != null) for (Map map : disassembly.getMaps())
 			{
